add test for follows and stocks
