+++ app/views/users/_follow_btn.html.erb
@@ -1,11 +1,11 @@
 <div id="follow-btn_<%= user.id %>">
   <% if !current_user?(user) %>
     <% if user.followed_by?(current_user) %>
-      <div class="chip chip_position white-text grey">
+      <div class="chip chip_position white-text grey btn_follow">
           <%= link_to 'フォロー中', user_relationships_path(user.id), method: :delete, class: "white-text", remote: true %>
       </div>
     <% else %>
-    <div class="chip chip_position white-text indigo">
+    <div class="chip chip_position white-text indigo btn_follow">
         <%= link_to 'フォローする', user_relationships_path(user.id), method: :post, class: "white-text", remote: true %>
     </div>
     <% end %>
