+++ spec/models/stock_spec.rb
@@ -3,5 +3,8 @@
 require 'rails_helper'
 
 RSpec.describe Stock, type: :model do
-  pending "add some examples to (or delete) #{__FILE__}"
+  it { is_expected.to validate_presence_of :user_id }
+  it { is_expected.to validate_presence_of :post_id }
+  it { is_expected.to belong_to(:user) }
+  it { is_expected.to belong_to(:post) }
 end
